Fix UserRole-s

+++ web/src/main/java/com/softserve/edu/controller/UserInfoDTO.java
@@ -11,9 +11,9 @@ public class UserInfoDTO {
     private String firstName;
     private String lastName;
     private String phone;
-    private Integer numberOfVerifications;
+    private Long numberOfVerifications;
 
-    public UserInfoDTO(String username, List<String> roles, String firstName, String lastName, String phone, Integer numberOfVerifications) {
+    public UserInfoDTO(String username, List<String> roles, String firstName, String lastName, String phone, Long numberOfVerifications) {
         this.username = username;
         this.roles = roles;
         this.firstName = firstName;
